# src/main/webapp/documentation/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a REST-CRUD-Application",
    "version": "1.0.0",
    "title": "REST-CRUD-Application",
    "termsOfService": "https://github.com/KiruxaUA/EPAM_7",
    "contact": {
      "email": "qwertyklopan@gmail.com"
    }
  },
  "host": "rest-crud-application.herokuapp.com",
  "basePath": "/api/v1",
  "tags": [
    {
      "name": "skill",
      "description": "Developer`s skill"
    },
    {
      "name": "account",
      "description": "Developer`s account"
    },
    {
      "name": "developer",
      "description": "Developer`s information"
    }
  ],
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/skills": {
      "get": {
        "tags": [
          "skill"
        ],
        "summary": "Get all skills",
        "description": "",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/Skill"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "skill"
        ],
        "summary": "Create new skill",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "skill",
            "schema": {
              "$ref": "#/definitions/Skill"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Cannot create skill"
          }
        }
      },
      "put": {
        "tags": [
          "skill"
        ],
        "summary": "Update an existing skill",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "skill",
            "schema": {
              "$ref": "#/definitions/Skill"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "204": {
            "description": "No Content"
          }
        }
      },
      "delete": {
        "tags": [
          "skill"
        ],
        "summary": "Delete skill",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Skill's ID",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Invalid parameter"
          },
          "404": {
            "description": "ID is not found"
          },
          "409": {
            "description": "Conflict occurred while deleting skill"
          }
        }
      }
    },
    "/skill": {
      "get": {
        "tags": [
          "skill"
        ],
        "summary": "Get skill by ID",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Skill's ID",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Skill"
            }
          },
          "400": {
            "description": "Invalid parameter ID"
          },
          "404": {
            "description": "ID is not found"
          }
        }
      }
    },
    "/accounts": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Get all accounts",
        "description": "",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/Account"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "account"
        ],
        "summary": "Create new account",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "account",
            "schema": {
              "$ref": "#/definitions/Account"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "409": {
            "description": "Conflict occurred while creating account"
          }
        }
      },
      "put": {
        "tags": [
          "account"
        ],
        "summary": "Update an existing account",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "account",
            "schema": {
              "$ref": "#/definitions/Account"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Cannot update account"
          },
          "404": {
            "description": "Account not found!"
          },
          "409": {
            "description": "Conflict occurred while updating account"
          }
        }
      },
      "delete": {
        "tags": [
          "account"
        ],
        "summary": "Delete account",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Account's ID",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Account not found!"
          },
          "409": {
            "description": "Conflict occurred while deleting account"
          }
        }
      }
    },
    "/account": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Get account's data",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Get account's ID",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Invalid account's ID"
          },
          "404": {
            "description": "Account's ID not found"
          }
        }
      }
    },
    "/developer": {
      "get": {
        "tags": [
          "developer"
        ],
        "summary": "Get developer's data",
        "description": "",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Developer"
            }
          },
          "400": {
            "description": "Invalid developer's ID"
          },
          "404": {
            "description": "Developer's ID not found"
          }
        }
      }
    },
    "/developers": {
      "get": {
        "tags": [
          "developer"
        ],
        "summary": "Get all developers",
        "description": "",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/Developer"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "developer"
        ],
        "summary": "Create new developer",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "developer",
            "schema": {
              "$ref": "#/definitions/Developer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "409": {
            "description": "Conflict occurred while creating developer"
          }
        }
      },
      "put": {
        "tags": [
          "developer"
        ],
        "summary": "Update an existing developer",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "developer",
            "schema": {
              "$ref": "#/definitions/Developer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Developer's ID not found!"
          },
          "409": {
            "description": "Conflict occurred while updating developer"
          }
        }
      },
      "delete": {
        "tags": [
          "developer"
        ],
        "summary": "Delete developer",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Developer's ID",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Developer's ID not found!"
          },
          "409": {
            "description": "Conflict occurred while deleting developer"
          }
        }
      }
    }
  },
  "definitions": {
    "Skill": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "Account": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "status": {
          "type": "string",
          "enum": [
            "ACTIVE",
            "BANNED",
            "DELETED"
          ]
        }
      }
    },
    "Developer": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "skills": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Skill"
          }
        },
        "account": {
          "$ref": "#/definitions/Account"
        }
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  }
}